Instantiate for all prime fields
